requirements.txt
pandas
numpy
scikit-learn
azure-functions
openpyxl
azure-storage-blob
python-dotenv